Add total CPU nums to monitor (#140)
* Add total CPU nums to monitor

* fix unittest keyerro

* yapf

* test_max_memory error, add more sleep time

* Add more sleep time in job._reply_client_heartbeat

* Try sleep 6 times to check actor number.

+++ parl/remote/tests/actor_max_memory_test.py
@@ -26,7 +26,7 @@ from parl.remote.monitor import ClusterMonitor
 from multiprocessing import Process
 
 
-@parl.remote_class(max_memory=300)
+@parl.remote_class(max_memory=350)
 class Actor(object):
     def __init__(self, x=10):
         self.x = x
@@ -64,10 +64,18 @@ class TestMaxMemory(unittest.TestCase):
         time.sleep(20)
         self.assertEqual(1, cluster_monitor.data['clients'][0]['actor_num'])
         del actor
-        time.sleep(5)
+        time.sleep(10)
         p = Process(target=self.actor)
         p.start()
-        time.sleep(30)
+
+        for _ in range(6):
+            x = cluster_monitor.data['clients'][0]['actor_num']
+            if x == 0:
+                break
+            else:
+                time.sleep(10)
+        if x == 1:
+            raise ValueError("Actor max memory test failed.")
         self.assertEqual(0, cluster_monitor.data['clients'][0]['actor_num'])
         p.terminate()
